move workflow utils into utils/workflow.js
